+++ AdventOfCode2019/Day07/AmplifierControllerSoftware.cs
@@ -11,10 +11,10 @@ namespace AdventOfCode2019.Day07
     {
         public IList<Amplifier> Amplifiers { get; set; }
 
-        public AmplifierControllerSoftware(IEnumerable<int> integers, IEnumerable<int> phaseSettings)
+        public AmplifierControllerSoftware(IEnumerable<long> integers, IEnumerable<int> phaseSettings)
         {
             var phasesArray = phaseSettings as int[] ?? phaseSettings.ToArray();
-            var integersArray = integers as int[] ?? integers.ToArray();
+            var integersArray = integers as long[] ?? integers.ToArray();
             if (phasesArray.Any(phase => phasesArray.Count(p => p == phase) > 1))
             {
                 throw new ArgumentException("Each phase can be only used once");
@@ -33,7 +33,7 @@ namespace AdventOfCode2019.Day07
         /// Runs all amplifiers and returns output of last amplifier.
         /// </summary>
         /// <returns>Output of last amplifier</returns>
-        public int GetThrusterSignal()
+        public long GetThrusterSignal()
         {
             for (var i = 0; i < Amplifiers.Count; i++)
             {
@@ -55,7 +55,7 @@ namespace AdventOfCode2019.Day07
         /// Loops them until they all stop. Then the last amplifier produces result.
         /// </summary>
         /// <returns>Result outputted by last amplifier after its last run</returns>
-        public int GetFeedbackLoopThrusterSignal()
+        public long GetFeedbackLoopThrusterSignal()
         {
             while (true)
             {
